@@ -10,21 +10,25 @@
   "dependencies": {
     "@codemirror/lang-html": "^6.4.5",
     "@date-io/luxon": "^3.0.0",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
     "@graphql-codegen/cli": "^4.0.0",
     "@graphql-codegen/typescript": "^4.0.0",
     "@graphql-codegen/typescript-graphql-request": "^5.0.0",
     "@graphql-codegen/typescript-operations": "^4.0.0",
     "@graphql-typed-document-node/core": "^3.2.0",
-    "@material-table/core": "^0.2.62",
+    "@hello-pangea/dnd": "^16.6.0",
+    "@material-table/core": "^6.4.4",
     "@mui/icons-material": "^5.15.18",
     "@mui/material": "^5.15.18",
-    "@mui/styles": "^5.15.18",
+    "@mui/system": "^5.15.15",
     "@mui/x-date-pickers": "^7.5.0",
     "@react-pdf-viewer/core": "^3.12.0",
     "@tinymce/tinymce-react": "^5.0.1",
-    "@uiw/react-codemirror": "^4.21.7",
+    "@types/mime": "^2.0.3",
+    "@types/tinymce": "^4.6.5",
+    "@types/yup": "^0.29.13",
+    "@uiw/react-codemirror": "^4.22.1",
     "@user-office-software/duo-localisation": "^1.2.0",
     "@user-office-software/duo-validation": "^5.0.2",
     "clsx": "^2.1.0",
@@ -32,8 +36,8 @@
     "graphql": "^16.8.1",
     "graphql-request": "^6.1.0",
     "html-react-parser": "^1.4.12",
-    "i18next": "^22.4.11",
-    "i18next-http-backend": "^2.2.0",
+    "i18next": "^23.11.5",
+    "i18next-http-backend": "^2.5.2",
     "immer": "^9.0.7",
     "jwt-decode": "^4.0.0",
     "luxon": "^3.4.4",
@@ -42,11 +46,10 @@
     "npm": "^10.5.2",
     "pdfjs-dist": "^3.4.120",
     "query-string": "^6.14.0",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-fast-compare": "^3.2.0",
-    "react-i18next": "^12.2.0",
+    "react-i18next": "^14.1.2",
     "react-router": "^5.3.1",
     "react-router-dom": "^5.3.1",
     "tinymce": "^7.1.0",
@@ -87,27 +90,26 @@
     "@types/luxon": "^3.4.2",
     "@types/mime": "^2.0.3",
     "@types/node": "^18.0.0",
-    "@types/react": "^17.0.44",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^17.0.16",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@types/react-router": "^5.1.18",
     "@types/react-router-dom": "^5.3.3",
-    "@types/react-test-renderer": "^17.0.1",
+    "@types/react-test-renderer": "^18.3.0",
     "@types/tinymce": "^4.6.9",
     "@types/yup": "^0.29.13",
     "@typescript-eslint/eslint-plugin": "^7.7.0",
     "@typescript-eslint/parser": "^7.7.0",
-    "@vitejs/plugin-react": "^4.2.0",
+    "@vitejs/plugin-react": "^4.3.0",
     "dotenv": "^16.4.5",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-unused-imports": "^3.1.0",
     "prettier": "3.2.4",
-    "react-test-renderer": "^17.0.0",
+    "react-test-renderer": "^18.3.1",
     "vite": "^5.0.13"
   },
   "engines": {
