@@ -3,7 +3,6 @@ import CloseIcon from '@mui/icons-material/Close';
 import Edit from '@mui/icons-material/Edit';
 import { IconButton } from '@mui/material';
 import Button from '@mui/material/Button';
-import makeStyles from '@mui/styles/makeStyles';
 import React, { SetStateAction, useState } from 'react';
 import {
   DecodedValueMap,
@@ -21,14 +20,6 @@ import InputDialog from 'components/common/InputDialog';
 import { setSortDirectionOnSortColumn } from 'utils/helperFunctions';
 import { tableIcons } from 'utils/materialIcons';
 import { FunctionType } from 'utils/utilTypes';
-
-const useStyles = makeStyles((theme) => ({
-  closeButton: {
-    position: 'absolute',
-    right: theme.spacing(1),
-    top: theme.spacing(1),
-  },
-}));
 
 export type UrlQueryParamsType = {
   search: QueryParamConfig<string | null | undefined>;
@@ -85,7 +76,6 @@ export function SuperMaterialTable<Entry extends EntryID>({
 }: MaterialTableProps<Entry> & SuperProps<Entry>) {
   const [show, setShow] = useState(false);
   const [editObject, setEditObject] = useState<Entry | null>(null);
-  const classes = useStyles();
 
   let { data, columns } = props;
   const {
@@ -205,7 +195,11 @@ export function SuperMaterialTable<Entry extends EntryID>({
       >
         <IconButton
           data-cy="close-modal-btn"
-          className={classes.closeButton}
+          sx={(theme) => ({
+            position: 'absolute',
+            right: theme.spacing(1),
+            top: theme.spacing(1),
+          })}
           onClick={() => {
             setEditObject(null);
             setShow(false);
