@@ -1,19 +1,6 @@
 import InputAdornment from '@mui/material/InputAdornment';
 import TextField, { TextFieldProps } from '@mui/material/TextField';
-import makeStyles from '@mui/styles/makeStyles';
-import clsx from 'clsx';
 import React, { ChangeEvent, useState } from 'react';
-
-const useStyles = makeStyles((theme) => ({
-  error: {
-    color: theme.palette.error.main,
-  },
-  adornmentPosition: {
-    alignSelf: 'flex-end',
-    alignItems: 'flex-end',
-    marginBottom: -theme.spacing(0.5),
-  },
-}));
 
 const TextFieldWithCounter = (
   props: TextFieldProps & {
@@ -21,7 +8,6 @@ const TextFieldWithCounter = (
     isCounterHidden?: boolean;
   }
 ) => {
-  const classes = useStyles();
   const [textLen, setTextLen] = useState(
     props.value ? (props.value as string).length : 0
   );
@@ -49,9 +35,14 @@ const TextFieldWithCounter = (
           endAdornment: !isCounterHidden && (
             <InputAdornment
               position="end"
-              className={clsx({
-                [classes.adornmentPosition]: other.multiline,
-                [classes.error]: maxLen && textLen > maxLen,
+              sx={(theme) => ({
+                ...(other.multiline && {
+                  alignSelf: 'flex-end',
+                  alignItems: 'flex-end',
+                  marginBottom: -theme.spacing(0.5),
+                }),
+                ...(maxLen &&
+                  textLen > maxLen && { color: theme.palette.error.main }),
               })}
             >
               {counter}
