@@ -1,58 +1,47 @@
+import Box, { BoxProps } from '@mui/material/Box';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
-import makeStyles from '@mui/styles/makeStyles';
-import PropTypes from 'prop-types';
 import React from 'react';
-
-const useStyles = makeStyles((theme) => ({
-  container: {
-    marginTop: '35px',
-    position: 'relative',
-  },
-  heading: {
-    top: '-15px',
-    left: '10px',
-    position: 'absolute',
-    marginBottom: '-15px',
-    background: '#FFF',
-    zIndex: 1,
-    padding: '0 8px',
-    color: theme.palette.grey[600],
-  },
-  contents: {
-    border: '1px solid #ccc',
-    borderRadius: '5px',
-    paddingTop: '26px',
-    padding: '16px',
-    zIndex: 0,
-  },
-}));
 
 type TitledContainerProps = {
   children: NonNullable<React.ReactNode>;
   label?: string;
-} & React.HTMLProps<HTMLDivElement>;
+} & BoxProps;
 
 const TitledContainer = (props: TitledContainerProps) => {
   const { children, label, ...rest } = props;
-  const classes = useStyles();
 
   return (
-    <div className={classes.container} {...rest}>
-      <Typography variant="h6" component="h2" className={classes.heading}>
+    <Box sx={{ marginTop: '35px', position: 'relative' }} {...rest}>
+      <Typography
+        variant="h6"
+        component="h2"
+        sx={(theme) => ({
+          top: '-15px',
+          left: '10px',
+          position: 'absolute',
+          marginBottom: '-15px',
+          background: '#FFF',
+          zIndex: 1,
+          padding: '0 8px',
+          color: theme.palette.grey[600],
+        })}
+      >
         {label}
       </Typography>
-      <Container className={classes.contents}>{children}</Container>
-    </div>
+      <Container
+        sx={{
+          border: '1px solid #ccc',
+          borderRadius: '5px',
+          paddingTop: '26px',
+          padding: '16px',
+          zIndex: 0,
+        }}
+      >
+        {children}
+      </Container>
+    </Box>
   );
-};
-
-TitledContainer.propTypes = {
-  label: PropTypes.string.isRequired,
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]).isRequired,
 };
 
 export default TitledContainer;
