@@ -9,7 +9,7 @@ import {
   DraggingStyle,
   Droppable,
   NotDraggingStyle,
-} from 'react-beautiful-dnd';
+} from '@hello-pangea/dnd';
 
 import { ProposalStatus } from 'generated/sdk';
 
