@@ -23,7 +23,7 @@ import {
   DraggingStyle,
   Droppable,
   NotDraggingStyle,
-} from 'react-beautiful-dnd';
+} from '@hello-pangea/dnd';
 import * as Yup from 'yup';
 
 import { getQuestionaryComponentDefinition } from 'components/questionary/QuestionaryComponentRegistry';
